+++ assets/js/editor-script-block.js
@@ -13,6 +13,7 @@
 	wp.blocks.unregisterBlockStyle( 'core/separator', 'wide' );
 	wp.blocks.unregisterBlockStyle( 'core/separator', 'dots' );
 	wp.blocks.unregisterBlockStyle( 'core/button', 'outline' );
+	wp.blocks.unregisterBlockStyle( 'core/quote', 'plain' );
 
 	/**
 	 * Handle embed block variations.