+++ footer.php
@@ -17,7 +17,7 @@
 			wp_nav_menu(
 				array(
 					'theme_location' => 'policy_menu',
-					'menu_class'     => 'list flex flex-wrap list-none my-3',
+					'menu_class'     => 'list flex flex-wrap list-none my-3 p-0',
 					'fallback_cb'    => false, // Do not fall back to wp_page_menu().
 				)
 			);