+++ inc/class/class-disclosure-nav-menu.php
@@ -45,7 +45,7 @@ class Disclosure_Nav_Menu extends Walker_Nav_Menu {
 		$output .= sprintf(
 			'<ul id="submenu-%s" class="%s" aria-label="%s">',
 			$this->cur_item->ID, // ID.
-			'disclosure-submenu level--' . ( $depth + 1 ) . ' ' . esc_attr( $level_class ) . ' list-none menu:absolute z-10 min-w-40 w-full pl-4 menu:pl-0 menu:w-auto has-black-background-color has-white-color menu:border hidden', // class.
+			'disclosure-submenu level--' . ( $depth + 1 ) . ' ' . esc_attr( $level_class ) . ' list-none menu:absolute z-10 min-w-40 w-full menu:w-auto m-0 pl-2 menu:p-4 menu:pl-0  has-black-background-color has-white-color menu:border hidden', // class.
 			sprintf(
 				/* translators: %s parent page title. */
 				esc_html( _x( 'Submenu for %s', 'menu UI: aria landmark', 'avidly-theme' ) ),
@@ -73,7 +73,7 @@ class Disclosure_Nav_Menu extends Walker_Nav_Menu {
 
 		$classes   = empty( $item->classes ) ? array() : (array) $item->classes;
 		$classes[] = 'menu-item-' . $item->ID;
-		$classes[] = 'relative flex justify-between items-center flex-wrap menu:flex-nowrap my-2 menu:mx-2 leading-tight'; // Style helpers.
+		$classes[] = 'relative flex justify-between items-center flex-wrap menu:flex-nowrap menu:whitespace-nowrap leading-tight menu:ml-4'; // Style helpers.
 
 		/**
 		 * Filter the arguments for a single nav menu item.
@@ -126,7 +126,7 @@ class Disclosure_Nav_Menu extends Walker_Nav_Menu {
 		$atts['target'] = ! empty( $item->target ) ? $item->target : '';
 		$atts['rel']    = ! empty( $item->xfn ) ? $item->xfn : '';
 		$atts['href']   = ! empty( $item->url ) ? $item->url : '';
-		
+
 		// Add the 'aria-current' attribute to current menu items.
 		if ( in_array( 'current-menu-item', $classes ) ) {
 			$atts['aria-current'] = 'page';
@@ -187,7 +187,7 @@ class Disclosure_Nav_Menu extends Walker_Nav_Menu {
 
 			$item_output .= sprintf(
 				'<button class="%s" aria-expanded="false" aria-controls="submenu-%s" aria-haspopup="true"><span class="screen-reader-text">%s</span></button>',
-				'disclosure-menu-toggle toggle-level--' . ( $depth + 1 ) . ' relative ml-1 w-8 h-8 text-center border-2 border-current',
+				'disclosure-menu-toggle toggle-level--' . ( $depth + 1 ) . ' relative ml-1 w-8 h-8 text-center border-2 border-current mr-4 menu:mr-0',
 				esc_attr( $this->cur_item->ID ),
 				sprintf(
 					/* translators: Submenu toggle button: %s parent page title. */